已完成

+++ database/migrations/2018_11_22_083703_create_items_table.php
@@ -24,7 +24,7 @@ class CreateitemsTable extends Migration
 			$table->integer('item_unitprice4');
 			$table->integer('item_unitprice5');
 			$table->string('item_unit');
-			$table->integer('item_taxtype');
+			$table->unsignedTinyInteger('item_taxtype');
 			$table->longText('item_image')->nullable($value = true);
             $table->timestamps();
 			$table->softDeletes();
